Fix missing images
For some unknown reason, the images optimized by FileOptimizer are not always
synchronized by iCloud correctly -- an image file exists in iCloud but missing
in Windows and cannot be synched. If I manually insert the iamge file in
Windows, iCloud creates a copy of the image with a numbered suffix; but the
original never gets synced correctly, thus missing the file.